+++ src/tools/genpolicy/src/mount_and_storage.rs
@@ -10,7 +10,7 @@ use crate::pod;
 use crate::policy;
 use crate::settings;
 use crate::volume;
-use crate::my_agent;
+use crate::agent;
 
 use log::debug;
 use std::ffi::OsString;
@@ -95,7 +95,7 @@ fn adjust_termination_path(mount: &mut policy::KataMount, yaml_container: &pod::
 pub fn get_mount_and_storage(
     settings: &settings::Settings,
     p_mounts: &mut Vec<policy::KataMount>,
-    storages: &mut Vec<my_agent::Storage>,
+    storages: &mut Vec<agent::Storage>,
     yaml_volume: &volume::Volume,
     yaml_mount: &pod::VolumeMount,
 ) {
@@ -124,7 +124,7 @@ pub fn get_mount_and_storage(
 fn get_empty_dir_mount_and_storage(
     settings: &settings::Settings,
     p_mounts: &mut Vec<policy::KataMount>,
-    storages: &mut Vec<my_agent::Storage>,
+    storages: &mut Vec<agent::Storage>,
     yaml_mount: &pod::VolumeMount,
     memory_medium: bool,
 ) {
@@ -137,7 +137,7 @@ fn get_empty_dir_mount_and_storage(
     debug!("Settings emptyDir: {:?}", settings_empty_dir);
 
     if yaml_mount.subPathExpr.is_none() {
-        storages.push(my_agent::Storage {
+        storages.push(agent::Storage {
             driver: settings_empty_dir.driver.clone(),
             driver_options: Vec::new(),
             source: settings_empty_dir.source.clone(),
@@ -232,7 +232,7 @@ fn get_host_path_mount(
 fn get_config_map_mount_and_storage(
     settings: &settings::Settings,
     p_mounts: &mut Vec<policy::KataMount>,
-    storages: &mut Vec<my_agent::Storage>,
+    storages: &mut Vec<agent::Storage>,
     yaml_mount: &pod::VolumeMount,
 ) {
     let settings_volumes = &settings.volumes;
@@ -247,7 +247,7 @@ fn get_config_map_mount_and_storage(
         let mount_path = Path::new(&yaml_mount.mountPath).file_name().unwrap();
         let mount_path_str = OsString::from(mount_path).into_string().unwrap();
 
-        storages.push(my_agent::Storage {
+        storages.push(agent::Storage {
             driver: settings_config_map.driver.clone(),
             driver_options: Vec::new(),
             source: format!("{}{}$", &settings_config_map.mount_source, &yaml_mount.name),
