+++ src/tools/genpolicy/src/yaml.rs
@@ -21,7 +21,7 @@ use crate::secret;
 use crate::settings;
 use crate::stateful_set;
 use crate::volume;
-use crate::my_agent;
+use crate::agent;
 
 use async_trait::async_trait;
 use core::fmt::Debug;
@@ -58,7 +58,7 @@ pub trait K8sResource {
     fn get_container_mounts_and_storages(
         &self,
         policy_mounts: &mut Vec<policy::KataMount>,
-        storages: &mut Vec<my_agent::Storage>,
+        storages: &mut Vec<agent::Storage>,
         container: &pod::Container,
         settings: &settings::Settings,
     );
@@ -232,7 +232,7 @@ pub async fn k8s_resource_init(spec: &mut pod::PodSpec, use_cache: bool) {
 
 pub fn get_container_mounts_and_storages(
     policy_mounts: &mut Vec<policy::KataMount>,
-    storages: &mut Vec<my_agent::Storage>,
+    storages: &mut Vec<agent::Storage>,
     container: &pod::Container,
     settings: &settings::Settings,
     volumes: &Vec<volume::Volume>,
